split out of draw-map : paths / collate / flows.
 Moved to :
utils/draw/flow-controls.js : flowButtonsSel, configurejQueryTooltip()
utils/draw/collate-paths.js : trace_alias, trace_adj
utils/flows.js : Flow()
utils/draw/collate-paths.js : collateStacks()
utils/draw/collate-paths.js : countPaths(), countPathsWithData()
collate-paths.js :
  aliasesUniqueName(), ensureFeatureIndex(), featureLookupName(),
  collateData(), collateFeatureClasses(), maInMaAG(), collateStacks1(),
  pathsUnique_log(), log_maamm(), log_ffaa(), mmaa2text(),
stacks-adj.js : collateAdjacentAxes(), log_adjAxes(),  log_adjAxes_a(), isAdjacent()
stacks.js : axisId2Name()
collate-paths.js :
  getAliased(), collateStacksA(), objPut(),
  aliasesText(), aliasText(),
  addPathsToCollation(), addPathsByReferenceToCollation(),
  storePath(), filterPaths(), selectCurrentAdjPaths(),
  collateFeatureMap(), concatAndUnique(), featureStackAxes(),
collate-paths.js : collateMagm()
flows-controls.js : flows_showControls()
app/utils/draw/flow-controls.js: updateSelections_flowControls() (split the flows part of updateSelections() to a new function)
flows-controls.js : Flow.prototype.ExportDataToDiv()

frontend/app/
 a1f0938 187048 Jul 10 21:30 components/draw-map.js
 975ae82   1086 Jul 10 21:30 components/draw/flow-controls.js
 f5dba64   3334 Jul 10 22:06 components/goto-ensembl.js
 94f1cdd   6772 Jul 10 21:30 services/data/flows-collate.js
 637aef6   2915 Jul 10 19:35 templates/components/panel/manage-view.hbs
 7b641f9  38634 Jul 10 22:58 utils/draw/collate-paths.js
 b2c01e2   6302 Jul 10 19:04 utils/draw/flow-controls.js
 e529834    979 Jul  9 20:38 utils/flows.js
 e8f2f2d   3723 Jul 10 18:18 utils/stacks-adj.js
 3ac4307  53219 Jul  9 15:42 utils/stacks.js
