+++ frontend/app/templates/components/draw/flow-controls.hbs
@@ -1,0 +1,1 @@
+{{!-- will migrate to here from app/utils/draw/flow-controls.js --}}
