+++ frontend/app/templates/components/panel/manage-view.hbs
@@ -71,6 +71,8 @@
   <div class="panel-body">
     <div class="drawing-controls vspace1">
       <div class="flowButtons"></div>
+      {{!-- probably migrate the content of flowButtons (d3) into flow-controls.hbs (this hbs is initially empty)  --}}
+      {{draw/flow-controls viewOptions=(hash isShowUnique pathColourScale showScaffoldMarkers showAsymmetricAliases )}}
     </div>
   </div>
 {{/elem/panel-container}}
